Use the module widget in slicer rather than creating a new one for test
After reloading just use the widget that is in the main gui.
This makes it easier to work with, rather than needing to deal
with a floating window.

TODO: For some reason this requires two reloads to start working - will need
to research that one a bit and fix it later.