Map new fiducials through current transform
This simplifies the process of adding new landmarks to a registration
in progress, since the new points will be in the right approximate
location as determined by the current transform.  That is, the new
information added by the pair of points should not cause a change
to the calculation of something like a thin plate spline except perhaps
adding extra weight in a region.