Centralize plugin management
This commit introduces the function 'registerRegistrationPlugin' providing
a simple way to register plugins. It avoid plugin maintainer to add
boiler plate code dealing with the 'slicer.modules.registrationPlugins'
dictionary and it makes the infrastructure more robust to code
re-organization.

It also display a warning when a plugin failed to be loaded. For example,
if Slicer is build without SimpleITK support, the following is displayed:

8<----8<----8<----8<----8<----8<----8<----8<----
[...]
Number of registered modules: 139
LandmarkRegistration: Failed to load 'RegistrationLib.LocalSimpleITKPlugin' plugin: No module named SimpleITK
Number of instantiated modules: 139
[...]
8<----8<----8<----8<----8<----8<----8<----8<----