BUG: Fix Qt5 crash when selecting registration plugin
Fixes https://issues.slicer.org/view.php?id=4632

Prior to this commit, switching between registration type or local refinement
led to a crash. Commenting out the deleteLater() call in RegistrationPlugin.destroy
function was sufficient to avoid the crash. Only with Slicer r27484 built
against Qt5 (it has a newer version of PythonQt)

This commit removes the explicit widget references that are not needed or
explicitly set them to None (e.g, hotUpdateButton in ThinPlatePlugin).

It seems that invalid references to widgets were explicitly kept around
despite of being invalidated by the call to "deleteLater()" in the destroy
function.