BUG: Program crashed when markers were close to the image border.
Refactored the logic behind the ROI selection. We now take a fixed
sized ROI centered on the marker point and interesect it with the
image to get the final ROI which is used in the registration. Also
modified ther registration initialization so that it aligns the fixed
and moving markers even if they are not in the center of the ROI
(which is what happens when they are close to the image border).